Move form component to root components folder
