+++ src/components/FormComponent/useForm.ts
@@ -1,7 +1,7 @@
 import { useCallback, useEffect, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 
-import { RoutesList } from '../../../../components/NavigationMenu/useNavigationMenu';
+import { RoutesList } from '../NavigationMenu/useNavigationMenu';
 
 import { FieldsList, FieldsObject, UseFormProps } from './types';
