+++ src/pages/Login/components/LoginComponent/useLogin.ts
@@ -1,11 +1,11 @@
 import { useState, useCallback, useMemo } from 'react';
 
-import { useForm } from '../FormComponent/useForm';
+import { useForm } from '../../../../components/FormComponent/useForm';
 
 import { authAPI } from '../../../../api/auth';
 
 import { InputTypeEnum } from '../../../../components/Input';
-import { FieldsList } from '../FormComponent/types';
+import { FieldsList } from '../../../../components/FormComponent/types';
 import { UseLoginComponentProps } from './types';
 
 const fieldsList = [FieldsList.login, FieldsList.password];
